+++ CHANGES.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## v0.2.7-beta.9
 ### Added
 - Added `IScope.ContainedScopes` which lets you check scopes contained within another scope.
 - Added `IVariable.CanBeAccessedIn(IScope)` which lets you check if the variable can be accessed within a given scope.